changing float to double for consistency

+++ src/quadcopter_pilot/modules/accel/MMA7455L/MMA7455L_hardware.c
@@ -7,7 +7,7 @@
 #define B 0
 
 // scale accel output (-127 .. 127) to (-8 .. 8)
-float scale(int16_t x) {
+double scale(int16_t x) {
     // y = mx + b 
     // m = (y2 - y1) / (x2 - x1) = (8 - -8) / (127 - -127) = 0.062992125984252
     // b = y1 - m * x1 = -8 - 0.062992125984252 * -127 = 0
@@ -19,19 +19,19 @@ void init_accel_hardware(){
 }
 
 /* Returns the scaled accelerometer data for X axis */
-float get_accel_x(){
+double get_accel_x(){
 	//TODO
     return scale (0);
 }
 
 /* Returns the scaled accelerometer data for Y axis */
-float get_accel_y(){
+double get_accel_y(){
 	//TODO
     return scale (0);
 }
 
 /* Returns the scaled accelerometer data for Z axis */
-float get_accel_z(){
+double get_accel_z(){
 	//TODO
     return scale (0);
 }
